Add localization config, and remove duplicate code
